Better groups
- There were issues with the dates, it was not sorted. Sorting it by date... (using the timestamp, faster/safer than parsing the string date)
- I tried having a different behavior for withdrawals that come from another account exchange, but it was not worth it ... and not changing that much actually. So I left the code within an "if False", but it wasn't working, just checking if it was worth it.

- TODO: Add a UT for the date thing

So there's still way too many transactions.
So I just bundled all the same transactions in the same transaction, and changed the date to be the same, so that cointracking will not get confused (hopefully)

that's one single transaction per year, per exchange/currency. To december 31st, and
ONLY for the old years.